+++ docs/changelog/changelog.md
@@ -20,3 +20,4 @@
 - [Fill and migrate required meta-data for generated PDF](https://dimpact.atlassian.net/browse/DATA-163)
 - [Validate zaaknummer length per Zaaktype before migration](https://dimpact.atlassian.net/browse/DATA-92)
 - [Global mapping & migration of Documentstatussen](https://dimpact.atlassian.net/browse/DATA-195)
+- [Migrate Besluit of Zaak](https://dimpact.atlassian.net/browse/DATA-192)